Extract PyO3 stream exports
Move the stream-facing Python functions into their own module so the
native library remains within the module-size limit. Preserve the
existing FFI boundary, registration, and scoped CodeScene exception.
